pyproject.toml
[tool.poetry]
name = "handon-fx"
version = "0.1.0"
description = ""
authors = ["osa9 <osanine9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.2"
pandas-datareader = "^0.10.0"
yfinance = "^0.2.3"
pynamodb = "^5.3.4"
boto3 = "^1.26.37"
uuid6 = "^2022.10.25"
"Mastodon.py" = "^1.8.0"
python-dotenv = "^0.21.0"
cryptography = "^39.0.0"
http-ece = "^1.1.0"
bokeh = "^3.0.3"
numpy = "^1.24.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
